Update helper-npm-update
caveat: there is a small risk, like the recent problem we found with x-dash and jest's vs eslint's use of acorn and peer dependencies, that this might cause an app to break. There is the risk that an app might break just with top-level too, the more that gets updated the more risk there is.