Install all the peer-dependencies with a single `npm install`
Previously the helper would install the peer-dependencies one at a time which can lead to a slower install times compared to installing all the peer-dependencies in a single `npm install`.

I came across this because the project I was working on had a `postinstall` npm run-script which was slow and it was being run after each individual peer-dependency was installed. Switching to install all the peer-dependencies in one go meant that the slow `postinstall` script was only run once.